Move common method implementations to base
