+++ src/main/java/com/uber/cadence/internal/worker/SuspendableWorkerBase.java
@@ -1,0 +1,76 @@
+/*
+ *  Copyright 2012-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ *  Modifications copyright (C) 2017 Uber Technologies, Inc.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not
+ *  use this file except in compliance with the License. A copy of the License is
+ *  located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file. This file is distributed on
+ *  an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ *  express or implied. See the License for the specific language governing
+ *  permissions and limitations under the License.
+ */
+
+package com.uber.cadence.internal.worker;
+
+import java.util.concurrent.TimeUnit;
+
+public abstract class SuspendableWorkerBase implements SuspendableWorker {
+  private SuspendableWorker poller = new NoopSuspendableWorker();
+
+  final void setPoller(SuspendableWorker poller) {
+    this.poller = poller;
+  }
+
+  @Override
+  public boolean isStarted() {
+    return poller.isStarted();
+  }
+
+  @Override
+  public boolean isShutdown() {
+    return poller.isShutdown();
+  }
+
+  @Override
+  public boolean isTerminated() {
+    return poller.isTerminated();
+  }
+
+  @Override
+  public void shutdown() {
+    poller.shutdown();
+  }
+
+  @Override
+  public void shutdownNow() {
+    poller.shutdownNow();
+  }
+
+  @Override
+  public void awaitTermination(long timeout, TimeUnit unit) {
+    if (!poller.isStarted()) {
+      return;
+    }
+    poller.awaitTermination(timeout, unit);
+  }
+
+  @Override
+  public void suspendPolling() {
+    poller.suspendPolling();
+  }
+
+  @Override
+  public void resumePolling() {
+    poller.resumePolling();
+  }
+
+  @Override
+  public boolean isSuspended() {
+    return poller.isSuspended();
+  }
+}
